docs on JsError, move to lib.rs
